Several updates: ts -> js, add audio page and component dependencies, add simple YAML page, update switcher accordingly, various local corrections and updates, etc.
