--- .prettierrc.js
@@ -1,1 +1,0 @@
-module.exports = require('eslint-config-mantine/.prettierrc.js');
