--- .storybook/main.js
@@ -1,11 +1,0 @@
-module.exports = {
-  stories: ['../**/*.story.mdx', '../**/*.story.@(js|jsx|ts|tsx)'],
-  addons: [
-    'storybook-dark-mode',
-    {
-      name: 'storybook-addon-turbo-build',
-      options: { optimizationLevel: 2 },
-    },
-  ],
-  framework: '@storybook/react',
-};
