+++ components/ColorSchemeToggle.jsx
@@ -1,0 +1,24 @@
+import { ActionIcon, Group, useMantineColorScheme } from "@mantine/core";
+import { IconSun, IconMoonStars } from "@tabler/icons";
+
+const ColorSchemeToggle = () => {
+    const { colorScheme, toggleColorScheme } = useMantineColorScheme();
+
+    return (
+        <Group position="center" mt="xl">
+            <ActionIcon
+                onClick={() => toggleColorScheme()}
+                size="xl"
+                sx={(theme) => ({
+                    backgroundColor:
+                        theme.colorScheme === "dark" ? theme.colors.dark[6] : theme.colors.gray[0],
+                    color: theme.colorScheme === "dark" ? theme.colors.yellow[4] : theme.colors.blue[6],
+                })}
+            >
+                {colorScheme === "dark" ? <IconSun size={20}/> : <IconMoonStars size={20}/>}
+            </ActionIcon>
+        </Group>
+    );
+}
+
+export default ColorSchemeToggle;
