+++ components/PageSwitcher.jsx
@@ -1,0 +1,36 @@
+import { useRouter } from "next/router";
+import { ActionIcon, Group } from "@mantine/core";
+import { IconTex, IconMarkdown, IconTable, IconList, IconGrain } from "@tabler/icons";
+
+const PageSwitcher = () => {
+    const router = useRouter();
+
+    const pages = {
+        "tex": <IconTex size={200} stroke={1.5} />,
+        "md": <IconMarkdown size={200} stroke={1.5} />,
+        "csv": <IconTable size={200} stroke={1.5} />,
+        "yaml": <IconList size={200} stroke={1.5} />,
+        "audio": <IconGrain size={200} stroke={1.5} />
+    }
+
+    return (
+        <Group position="center" mt="xl" grow>
+            {Object.entries(pages).map(([key, value], i) =>
+                <ActionIcon
+                    key={i}
+                    onClick={() => router.push(`/${key}`)}
+                    size={240}
+                    sx={(theme) => ({
+                    backgroundColor:
+                        theme.colorScheme === "dark" ? theme.colors.dark[6] : theme.colors.light,
+                        color: theme.colors.yellow[4]
+                    })}
+                >
+                    {value}
+                </ActionIcon>)
+            }
+        </Group>
+    );
+}
+
+export default PageSwitcher;
