+++ package.json
@@ -1,22 +1,13 @@
 {
-  "name": "mantine-next-template",
+  "name": "htpaaiab",
   "version": "1.0.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build && next export",
-    "analyze": "ANALYZE=true next build",
+    "build": "next build",
+    "export": "next export",
     "start": "next start",
-    "typecheck": "tsc --noEmit",
-    "export": "next build && next export",
-    "lint": "next lint",
-    "jest": "jest",
-    "jest:watch": "jest --watch",
-    "prettier:check": "prettier --check \"**/*.{ts,tsx}\"",
-    "prettier:write": "prettier --write \"**/*.{ts,tsx}\"",
-    "test": "npm run prettier:check && npm run lint && npm run typecheck && npm run jest",
-    "storybook": "start-storybook -p 7001",
-    "storybook:build": "build-storybook"
+    "lint": "next lint"
   },
   "dependencies": {
     "@emotion/react": "^11.10.0",
@@ -24,15 +15,15 @@
     "@mantine/core": "5.8.3",
     "@mantine/hooks": "5.8.3",
     "@mantine/next": "5.8.3",
-    "@react-three/drei": "^9.56.4",
-    "@react-three/fiber": "^8.10.1",
+    "@mantine/prism": "^5.10.2",
     "@tabler/icons": "^1.113.0",
-    "@types/papaparse": "^5.3.7",
     "@unified-latex/unified-latex": "^1.2.2",
     "@unified-latex/unified-latex-util-ligatures": "^1.2.2",
+    "axios": "^1.3.1",
     "better-react-mathjax": "^2.0.2",
     "bibtex": "^0.9.0",
     "cookies-next": "^2.1.1",
+    "cors": "^2.8.5",
     "file-loader": "^6.2.0",
     "global": "^4.4.0",
     "js-yaml": "^4.1.0",
@@ -46,40 +37,12 @@
     "react-markdown": "^8.0.5",
     "react-router-dom": "^6.4.4",
     "react-social-media-embed": "^2.2.13",
-    "rehype-raw": "^6.1.1",
-    "save": "^2.9.0",
-    "three": "^0.149.0"
+    "rehype-raw": "^6.1.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.8",
-    "@next/eslint-plugin-next": "^12.1.4",
-    "@storybook/react": "^6.5.13",
-    "@testing-library/dom": "^8.12.0",
-    "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^13.0.0",
-    "@testing-library/user-event": "^14.0.4",
-    "@types/jest": "^27.4.1",
-    "@types/node": "^18.11.4",
-    "@types/react": "18.0.21",
-    "@typescript-eslint/eslint-plugin": "^5.30.0",
-    "@typescript-eslint/parser": "^5.30.0",
-    "babel-loader": "^8.2.4",
+    "@types/react": "18.0.27",
     "eslint": "^8.18.0",
-    "eslint-config-airbnb": "19.0.4",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-mantine": "2.0.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^26.1.1",
-    "eslint-plugin-jsx-a11y": "^6.6.0",
-    "eslint-plugin-react": "^7.30.1",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-storybook": "^0.5.7",
-    "eslint-plugin-testing-library": "^5.2.0",
-    "jest": "^27.5.1",
-    "prettier": "^2.7.1",
-    "storybook-addon-turbo-build": "^1.1.0",
-    "storybook-dark-mode": "^1.1.2",
-    "ts-jest": "^27.1.4",
-    "typescript": "4.8.4"
+    "eslint-config-next": "^13.1.6",
+    "typescript": "4.9.5"
   }
 }
