--- pages/_document.tsx
@@ -1,8 +1,0 @@
-import Document from 'next/document';
-import { createGetInitialProps } from '@mantine/next';
-
-const getInitialProps = createGetInitialProps();
-
-export default class _Document extends Document {
-    static getInitialProps = getInitialProps;
-}
