--- pages/additional.d.ts
@@ -1,19 +1,0 @@
-declare module '*.md' {
-    const content: string;
-    export default content;
-}
-
-declare module '*.csv' {
-    const content: string;
-    export default content;
-}
-
-declare module '*.tex' {
-    const content: string;
-    export default content;
-}
-
-declare module '*.yaml' {
-    const content: any;
-    export default content;
-}
