+++ pages/data.jsx
@@ -1,0 +1,10 @@
+
+const Data = () => {
+    return (
+        <>
+
+        </>
+    );
+}
+
+export default Data;
